Increased size of nbr in ServoPin_t to enable usage of alternative Pin Mappings
e.g. servo.attach(PB0 | ALT1) would fail, also the wrong pin number would be given to HardwareTimer.setMode() later

+++ src/PRDC_ServoHT.h
@@ -41,7 +41,7 @@
 #define INVALID_SERVO 255
 
 typedef struct  {
-  uint8_t nbr;            // a pin number from 0 to 255
+  uint16_t nbr;            // 16 bit pin number to enable alternative pin mappings
   uint8_t isActive;       // true if this channel is enabled, pin not pulsed if false
   uint8_t prevAttached;       // true if this channel is previusly attached
 } ServoPin_t;
